Getting a very strange problem.
Due to this:
https://github.com/facebook/create-react-app/blob/master/packages/babel-preset-react-app/create.js#L181

The path to generator are all absolute, resulting in the dist cannot be
used as a package.

Due to babel and tsc differs in how they compile the source, and for
some reason babel out put does not respect baseUrl, the solution for now
is thus:
+ Consume @roshub-design-system/dist/... instead
+ Remove baseUrl from design-system for now and reinvestigate in the
  future...
+ Add a postinstall script that build the dist
Dark time ahead for a design system that can work inside a monorepo...

OR, move it out of the mono-repo and link manually.